requirements.txt
pydantic==2.10.6
fastapi==0.115.8
uvicorn==0.34.0
openai==1.68.2
pyyaml==6.0.2
httpx==0.28.1
pytest==8.3.4
openenv-core>=0.0.1
